enum 타입정리
