+++ public/static/requireJudge/static.ts
@@ -1,3 +1,5 @@
+import { DEPOSIT_STATUS, POSITION_CODE, PROJECT_REQUIRE_JOIN_STATUS, PROJECT_REQUIRE_JUDGE_PROGRESS_STATUS, PROJECT_STATUS } from "public/lib/enum";
+
 export const PROJECT_REQUIRE_JUDGE_OPTION = [
   {
     label: '전체',
@@ -26,24 +28,24 @@ export const PROJECT_REQUIRE_JUDGE_OPTION = [
  */
 export const POSITION_CODE_ARRAY = [
   {
-    id: 30,
-    name: '디자인',
+    value: POSITION_CODE.DESIGN,
+    label: '디자인',
   },
   {
-    id: 14,
-    name: '프론트',
+    value: POSITION_CODE.FRONT,
+    label: '프론트',
   },
   {
-    id: 15,
-    name: '백엔드',
+    value: POSITION_CODE.BACKEND,
+    label: '백엔드',
   },
   {
-    id: 16,
-    name: '서버/인프라',
+    value: POSITION_CODE.SERVER,
+    label: '서버/인프라',
   },
   {
-    id: 64,
-    name: '기타',
+    value: POSITION_CODE.ETC,
+    label: '기타',
   },
 ];
 
@@ -52,48 +54,48 @@ export const POSITION_CODE_ARRAY = [
  */
 export const PROJECT_STATUS_ARRAY = [
   {
-    id: '19',
-    name: '모집중',
+    value: PROJECT_STATUS.RECRUITING,
+    label: '모집중',
   },
   {
-    id: '20',
-    name: '모집완료',
+    value: PROJECT_STATUS.RECRUITMENT_COMPLETED,
+    label: '모집완료',
   },
   {
-    id: '21',
-    name: '요구사항 심사중',
+    value: PROJECT_STATUS.REQUIREMENT_ASSESSMENT_IN_PROGRESS,
+    label: '요구사항 심사중',
   },
   {
-    id: '22',
-    name: '요구사항 반려',
+    value: PROJECT_STATUS.REQUIREMENT_ASSESSMENT_REJECTED,
+    label: '요구사항 반려',
   },
   {
-    id: '23',
-    name: '진행 대기',
+    value: PROJECT_STATUS.WAITING_TO_START,
+    label: '진행 대기',
   },
   {
-    id: '24',
-    name: '진행중',
+    value: PROJECT_STATUS.IN_PROGRESS,
+    label: '진행중',
   },
   {
-    id: '25',
-    name: '중도 종결',
+    value: PROJECT_STATUS.PARTIAL_TERMINATION,
+    label: '중도 종결',
   },
   {
-    id: '26',
-    name: '제출완료',
+    value: PROJECT_STATUS.SUBMISSION_COMPLETED,
+    label: '제출완료',
   },
   {
-    id: '27',
-    name: '심사중',
+    value: PROJECT_STATUS.ASSESSMENT_IN_PROGRESS,
+    label: '심사중',
   },
   {
-    id: '28',
-    name: '심사완료',
+    value: PROJECT_STATUS.ASSESSMENT_COMPLETED,
+    label: '심사완료',
   },
   {
-    id: '41',
-    name: '입금 기간 초과',
+    value: PROJECT_STATUS.DEPOSIT_PERIOD_EXPIRED,
+    label: '입금 기간 초과',
   },
 ];
 
@@ -102,20 +104,20 @@ export const PROJECT_STATUS_ARRAY = [
  */
 export const PROJECT_REQUIRE_JUDGE_PROGRESS_STATUS_ARRAY = [
   {
-    id: '32',
-    name: '요구사항 정의 중',
+    value: PROJECT_REQUIRE_JUDGE_PROGRESS_STATUS.REQUIREMENT_DEFINITION_IN_PROGRESS,
+    label: '요구사항 정의 중',
   },
   {
-    id: '33',
-    name: '대기',
+    value: PROJECT_REQUIRE_JUDGE_PROGRESS_STATUS.WAITING,
+    label: '대기',
   },
   {
-    id: '34',
-    name: '반려',
+    value: PROJECT_REQUIRE_JUDGE_PROGRESS_STATUS.REJECTED,
+    label: '반려',
   },
   {
-    id: '35',
-    name: '승인',
+    value: PROJECT_REQUIRE_JUDGE_PROGRESS_STATUS.APPROVED,
+    label: '승인',
   },
 ];
 
@@ -124,20 +126,20 @@ export const PROJECT_REQUIRE_JUDGE_PROGRESS_STATUS_ARRAY = [
  */
 export const PROJECT_REQUIRE_JOIN_STATUS_ARRAY = [
   {
-    id: '37',
-    name: '대기',
+    value: PROJECT_REQUIRE_JOIN_STATUS.WAITING,
+    label: '대기',
   },
   {
-    id: '38',
-    name: '반려',
+    value: PROJECT_REQUIRE_JOIN_STATUS.REJECTED,
+    label: '반려',
   },
   {
-    id: '39',
-    name: '승인',
+    value: PROJECT_REQUIRE_JOIN_STATUS.APPROVED,
+    label: '승인',
   },
   {
-    id: '40',
-    name: '취소',
+    value: PROJECT_REQUIRE_JOIN_STATUS.CANCELED,
+    label: '취소',
   },
 ];
 
@@ -146,88 +148,24 @@ export const PROJECT_REQUIRE_JOIN_STATUS_ARRAY = [
  */
 export const DEPOSIT_STATUS_ARRAY = [
   {
-    id: '43',
-    name: '대기',
+    value: DEPOSIT_STATUS.WAITING,
+    label: '대기',
   },
   {
-    id: '44',
-    name: '기간 초과',
+    value: DEPOSIT_STATUS.EXPIRED_PERIOD,
+    label: '기간 초과',
   },
   {
-    id: '45',
-    name: '입금 확인',
+    value: DEPOSIT_STATUS.DEPOSIT_CONFIRMED,
+    label: '입금 확인',
   },
   {
-    id: '46',
-    name: '환급 대기',
+    value: DEPOSIT_STATUS.REFUND_WAITING,
+    label: '환급 대기',
   },
   {
-    id: '47',
-    name: '환급 완료',
+    value: DEPOSIT_STATUS.REFUND_COMPLETED,
+    label: '환급 완료',
   },
 ];
 
-/**
- * [스킬 목록]
- */
-export const SKILLS = [
-  { id: 1, name: 'React', imageName: 'react' },
-  { id: 2, name: 'TypeScript', imageName: 'typeScript' },
-  { id: 3, name: 'JavaScript', imageName: 'js' },
-  { id: 4, name: 'Vue', imageName: 'vue' },
-  { id: 5, name: 'Next.js', imageName: 'nextjs' },
-  { id: 6, name: 'Svelte', imageName: 'svelte' },
-  { id: 7, name: 'Java', imageName: 'java' },
-  { id: 8, name: 'Spring', imageName: 'spring' },
-  { id: 9, name: 'Node.js', imageName: 'node_js' },
-  { id: 10, name: 'Nest.js', imageName: 'nestjs' },
-  { id: 11, name: 'Go', imageName: 'go' },
-  { id: 12, name: 'Kotlin', imageName: 'jb_kotlin' },
-  { id: 13, name: 'Express', imageName: 'express' },
-  { id: 14, name: 'MySQL', imageName: 'mySQL' },
-  { id: 15, name: 'MongoDB', imageName: 'mongoDB' },
-  { id: 16, name: 'Python', imageName: 'python' },
-  { id: 17, name: 'DJango', imageName: 'django' },
-  { id: 18, name: 'PHP', imageName: 'php' },
-  { id: 19, name: 'GraphQL', imageName: 'graphQL' },
-  { id: 20, name: 'Firebase', imageName: 'firebase' },
-  { id: 21, name: 'Flutter', imageName: 'flutter' },
-  { id: 22, name: 'Swift', imageName: 'swift' },
-  { id: 23, name: 'Unity', imageName: 'unity' },
-  { id: 24, name: 'AWS', imageName: 'AWS' },
-  { id: 25, name: 'Kubernetes', imageName: 'kubernetes' },
-  { id: 26, name: 'Docker', imageName: 'docker' },
-  { id: 27, name: 'Git', imageName: 'git' },
-  { id: 28, name: 'Figma', imageName: 'figma' },
-  { id: 29, name: 'Zeplin', imageName: 'zeplin' },
-  { id: 30, name: 'Jest', imageName: 'jest' },
-  { id: 31, name: 'C', imageName: 'c' },
-  { id: 32, name: 'CSS', imageName: 'css_3' },
-  { id: 33, name: 'HTML', imageName: 'html_5' },
-  { id: 34, name: 'Bit Bucket', imageName: 'bitbucket' },
-  { id: 35, name: 'JQuery', imageName: 'jquery' },
-  { id: 36, name: 'C++', imageName: 'c++' },
-  { id: 38, name: 'R', imageName: 'r_lang' },
-  { id: 39, name: 'C#', imageName: 'c#' },
-  { id: 40, name: 'Ruby', imageName: 'ruby' },
-  { id: 42, name: 'Angular', imageName: 'angular' },
-  { id: 43, name: 'Drupal', imageName: 'drupal' },
-  { id: 44, name: 'Joomla', imageName: 'joomla' },
-  { id: 45, name: 'Wordpress', imageName: 'wordpress' },
-  { id: 46, name: 'Redux', imageName: 'redux' },
-  { id: 47, name: 'Weebly', imageName: 'weebly' },
-  { id: 48, name: 'Yii', imageName: 'yii 1' },
-  { id: 49, name: 'Dreamweaver', imageName: 'adobe_dreamweaver' },
-  { id: 50, name: 'Dribbble', imageName: 'dribbble' },
-  { id: 51, name: 'Behance', imageName: 'behance' },
-  { id: 52, name: 'Invision', imageName: 'invision' },
-  { id: 53, name: 'Autodesk', imageName: 'autodesk' },
-  { id: 54, name: 'Lightroom', imageName: 'adobe_lightroom' },
-  { id: 55, name: 'Illustrator', imageName: 'adobe_illustrator' },
-  { id: 56, name: 'After Effects', imageName: 'adobe_after_effects' },
-  { id: 57, name: 'Photoshop', imageName: 'adobe_photoshop' },
-  { id: 58, name: 'Indesign', imageName: 'adobe_indesign' },
-  { id: 59, name: 'Xd', imageName: 'adobe_xd' },
-  { id: 60, name: 'Sketch', imageName: 'sketch' },
-  { id: 61, name: 'Framer', imageName: 'framer' },
-];
