+++ src/components/hooks/useInfoMessage.tsx
@@ -14,7 +14,7 @@ const useInfoMessage = <T,>(defaultInfoMessage: TYPE<T>) => {
   const infoMessagesKeys = Object.keys(defaultInfoMessage) as [keyof T];
 
   const onChangeInfoMessage = <U,>(value: string, name: keyof T & string, params: U) => {
-    const message = value?.length === 0 ? '' : defaultInfoMessage[name].validate(value, params, 'password').children;
+    const message = value?.length === 0 ? '' : defaultInfoMessage[name].validate(value, params, name).children;
     const isValidate = defaultInfoMessage[name].validate(value, params).status;
